+++ cfdmod/use_cases/pressure/chunking.py
@@ -1,44 +1,50 @@
 import math
 import pathlib
-from typing import Callable
+from typing import Callable, ClassVar
 
 import numpy as np
 import pandas as pd
 from lnas import LnasGeometry
 
-from cfdmod.use_cases.pressure.extreme_values import ExtremeValuesParameters
-from cfdmod.use_cases.pressure.statistics import Statistics
+from cfdmod.use_cases.pressure.statistics import BasicStatisticModel, ParameterizedStatisticModel
 from cfdmod.use_cases.pressure.zoning.processing import calculate_statistics
+from cfdmod.utils import convert_matrix_into_dataframe
 
 
 class HDFGroupInterface:
-    # HDF keys follow the convention step_{formatted initial_step}_group_{formatted group_idx}
+    # HDF keys follow the convention /step_{formatted initial_step}_group_{formatted group_idx}
     # Step information comes from simulation results
+    TEMPORAL_PREFIX: ClassVar[str] = "/step"
+    GROUP_PREFIX: ClassVar[str] = "group"
+
+    @classmethod
+    def time_key(cls, initial_time: float) -> str:
+        return f"{cls.TEMPORAL_PREFIX}{int(initial_time):07}"
 
     @classmethod
     def get_point_group_key(cls, timestep_group_lbl: str, group_idx: int) -> str:
-        return timestep_group_lbl + f"_group_{group_idx:03}"
+        return "_".join([timestep_group_lbl, cls.GROUP_PREFIX + f"{group_idx:04}"])
 
     @classmethod
     def get_available_point_groups(cls, hdf_keys: list[str]) -> set[str]:
-        return set(["group" + k.split("group")[1] for k in hdf_keys])
+        return set([cls.GROUP_PREFIX + k.split(cls.GROUP_PREFIX)[1] for k in hdf_keys])
 
     @classmethod
     def get_timestep_keys_for_group(cls, hdf_keys: list[str], group_key: str) -> list[str]:
-        return [k for k in hdf_keys if k.split("group")[1] in group_key]
+        return [k for k in hdf_keys if k.split(cls.GROUP_PREFIX)[1] in group_key]
 
     @classmethod
     def filter_groups(
         cls, group_keys: list[str], timestep_range: tuple[float, float]
     ) -> list[str]:
-        steps = [int(key.split("step")[1]) for key in group_keys]
+        steps = [int(key.replace(cls.TEMPORAL_PREFIX, "")) for key in group_keys]
         lower_values = [x for x in steps if x <= timestep_range[0]]
         if len(lower_values) == 0:
-            return [f"/step{step:07}" for step in steps if step <= timestep_range[1]]
+            return [cls.time_key(step) for step in steps if step <= timestep_range[1]]
         else:
             initial_step = max(lower_values)
             return [
-                f"/step{step:07}" for step in steps if initial_step <= step <= timestep_range[1]
+                cls.time_key(step) for step in steps if initial_step <= step <= timestep_range[1]
             ]
 
 
@@ -66,28 +72,28 @@ def split_into_chunks(
         raise ValueError("There must be at least two steps in each chunk")
 
     for i in range(number_of_chunks):
-        min_step, max_step = i * step, min((i + 1) * step - 1, len(time_arr) - 1)
+        initial_step, end_step = i * step, min((i + 1) * step - 1, len(time_arr) - 1)
         df: pd.DataFrame = time_series_df.loc[
-            (time_series_df.time_step >= time_arr[min_step])
-            & (time_series_df.time_step <= time_arr[max_step])
+            (time_series_df.time_step >= time_arr[initial_step])
+            & (time_series_df.time_step <= time_arr[end_step])
         ].copy()
 
-        range_lbl = f"range_{int(time_arr[min_step])}_{int(time_arr[max_step])}"
+        range_lbl = HDFGroupInterface.time_key(initial_time=time_arr[initial_step])
 
-        df.to_hdf(path_or_buf=output_path, key=range_lbl, mode="a", index=False, format="t")
+        df.to_hdf(path_or_buf=output_path, key=range_lbl, mode="a", index=False, format="table")
 
 
 def calculate_statistics_for_groups(
     grouped_data_path: pathlib.Path,
-    statistics: list[Statistics],
-    extreme_params: ExtremeValuesParameters | None,
+    statistics: list[BasicStatisticModel | ParameterizedStatisticModel],
+    time_scale_factor: float,
 ) -> pd.DataFrame:
     """Calculates statistics for groups of points
 
     Args:
         grouped_data_path (pathlib.Path): Path of grouped data (HDF)
-        statistics (list[Statistics]): List of statistics to apply
-        extreme_params (ExtremeValuesParameters | None): Parameters for extreme values analysis
+        statistics (list[BasicStatisticModel | ParameterizedStatisticModel]): List of statistics with parameters to apply
+        time_scale_factor (float): Factor for converting time scales from CST values
 
     Returns:
         pd.DataFrame: Statistics dataframe
@@ -104,21 +110,20 @@ def calculate_statistics_for_groups(
             for key in keys_for_group:
                 df = groups_store.get(key)
                 group_dfs.append(df)
-
-            cp_data = pd.concat(group_dfs).sort_values(by=["time_step", "point_idx"])
+            cp_data = pd.concat(group_dfs).sort_values(by=["time_step"])
             cp_stats = calculate_statistics(
                 cp_data,
                 statistics_to_apply=statistics,
-                variables=["cp"],
-                group_by_key="point_idx",
-                extreme_params=extreme_params,
+                time_scale_factor=time_scale_factor,
             )
             del cp_data
             stats_df.append(cp_stats)
 
-    full_stats = pd.concat(stats_df).sort_values(by=["point_idx"])
+    full_stats = pd.concat(stats_df).T
+    full_stats.reset_index(inplace=True)
+    full_stats.rename(columns={"index": "scalar"}, inplace=True)
 
-    return full_stats
+    return full_stats[["scalar"] + sorted([col for col in full_stats.columns if col != "scalar"])]
 
 
 def divide_timeseries_in_groups(
@@ -137,17 +142,19 @@ def divide_timeseries_in_groups(
 
         for group_lbl in groups:
             coefficient_data = data_store.get(group_lbl)
-            if pt_groups == None:
-                points_arr = coefficient_data.point_idx.unique()
+            if pt_groups is None:
+                points_arr = np.array(
+                    [col for col in coefficient_data.columns if col != "time_step"], dtype=np.int32
+                )
                 n_per_group = len(points_arr) // n_groups
                 pt_groups = np.split(points_arr, range(n_per_group, len(points_arr), n_per_group))
 
             for i, points_in_group in enumerate(pt_groups):
-                group_data = coefficient_data.loc[
-                    coefficient_data.point_idx.isin(points_in_group)
-                ].copy()
+                group_data = coefficient_data[points_in_group].copy()
+                group_data["time_step"] = coefficient_data["time_step"]
                 group_key = HDFGroupInterface.get_point_group_key(group_lbl, i)
-                group_data.to_hdf(output_path, key=group_key, mode="a")
+                group_data.to_hdf(output_path, key=group_key, mode="a", format="table")
+                del group_data
 
 
 def process_timestep_groups(
@@ -155,6 +162,7 @@ def process_timestep_groups(
     geometry_df: pd.DataFrame,
     geometry: LnasGeometry,
     processing_function: Callable[[pd.DataFrame, pd.DataFrame, LnasGeometry], pd.DataFrame],
+    data_label: str = "cp",
 ) -> pd.DataFrame:
     """Process the timestep groups with geometric properties
 
@@ -164,6 +172,7 @@ def process_timestep_groups(
         geometry (LnasGeometry): Geometry to be processed. Needed for evaluating representative area and volume
         processing_function (Callable[[pd.DataFrame, pd.DataFrame, LnasGeometry], pd.DataFrame]):
             Coefficient processing function
+        data_label (str): Label of the tabulated time series dataframe. Defaults to "cp".
 
     Returns:
         pd.DataFrame: Transformed pressure coefficient time series
@@ -175,6 +184,10 @@ def process_timestep_groups(
 
         for store_group in store_groups:
             sample = df_store.get(store_group)
+            if "point_idx" not in sample.columns:
+                # If point_idx is not in dataframe columns, then matrix form is assumed
+                # and needs to be converted to older format
+                sample = convert_matrix_into_dataframe(sample, value_data_label=data_label)
             coefficient_data = processing_function(sample, geometry_df, geometry)
             processed_samples.append(coefficient_data)