+++ cfdmod/use_cases/pressure/cp_data.py
@@ -13,9 +13,8 @@ from cfdmod.use_cases.pressure.chunking import (
     divide_timeseries_in_groups,
 )
 from cfdmod.use_cases.pressure.cp_config import CpConfig
-from cfdmod.use_cases.pressure.extreme_values import ExtremeValuesParameters
 from cfdmod.use_cases.pressure.path_manager import CpPathManager
-from cfdmod.utils import create_folders_for_file
+from cfdmod.utils import convert_dataframe_into_matrix, create_folders_for_file
 
 
 def transform_to_cp(
@@ -37,23 +36,19 @@ def transform_to_cp(
     Returns:
         pd.DataFrame: Dataframe of pressure coefficient data for the body
     """
-    average_static_pressure = press_data["rho"].to_numpy().mean()
+    static_pressure_array = press_data[0].to_numpy()
+    average_static_pressure = static_pressure_array.mean()
     dynamic_pressure = 0.5 * average_static_pressure * (reference_vel * correction_factor) ** 2
     cs_square = 1 / 3
     multiplier = cs_square / dynamic_pressure
 
-    df_pressure = press_data.set_index("time_step")
-    df_body = body_data.set_index("time_step")
+    press = static_pressure_array if ref_press_mode == "instantaneous" else average_static_pressure
+    convert_to_cp = lambda col: multiplier * (col - press)
+    df_cp = body_data.apply(
+        lambda col: col if col.name == "time_step" else col.apply(convert_to_cp)
+    )
 
-    if ref_press_mode == "instantaneous":
-        df_body["cp"] = multiplier * (df_body["rho"] - df_body.index.map(df_pressure["rho"]))
-    elif ref_press_mode == "average":
-        df_body["cp"] = multiplier * (df_body["rho"] - average_static_pressure)
-
-    df_body.reset_index(inplace=True)
-    df_body.drop(columns=["rho"], inplace=True)
-
-    return df_body
+    return df_cp
 
 
 def filter_data(data: pd.DataFrame, timestep_range: tuple[float, float]) -> pd.DataFrame:
@@ -107,17 +102,41 @@ def process_raw_groups(
                 keys_to_include = HDFGroupInterface.filter_groups(
                     body_groups, cp_config.timestep_range
                 )
+
+            average_value = None
+
+            if cp_config.reference_pressure == "average":
+                static_dfs = []
+                for store_group in static_groups:
+                    static_dfs.append(static_store.get(store_group))
+                merged_df = pd.concat(static_dfs)
+                # Old versions index the column with rho and new versions use point index (0)
+                # to label the column. Hence the condition below
+                average_value = (
+                    merged_df["rho"].mean() if "rho" in merged_df.columns else merged_df[0].mean()
+                )
+
             for store_group in body_groups:
-                if more_than_one_group:
-                    if store_group not in keys_to_include:
-                        continue
+                if more_than_one_group and store_group not in keys_to_include:
+                    continue
 
                 static_df = static_store.get(store_group)
                 static_df = filter_data(static_df, timestep_range=cp_config.timestep_range)
                 body_df = body_store.get(store_group)
                 body_df = filter_data(body_df, timestep_range=cp_config.timestep_range)
 
-                if (static_df.time_step.unique() != body_df.time_step.unique()).all():
+                # This logic should be removed in later versions
+                if "point_idx" in body_df.columns:
+                    # Data is in older format, must convert to matrix
+                    body_df = convert_dataframe_into_matrix(body_df)
+                if "point_idx" in static_df.columns:
+                    # Data is in older format, must convert to matrix
+                    static_df = convert_dataframe_into_matrix(static_df)
+
+                if average_value != None:
+                    static_df[0] = average_value
+
+                if any(static_df.time_step.unique() != body_df.time_step.unique()):
                     raise Exception(f"Timesteps for key {store_group} do not match!")
 
                 coefficient_data = transform_to_cp(
@@ -127,7 +146,7 @@ def process_raw_groups(
                     ref_press_mode=cp_config.reference_pressure,
                     correction_factor=cp_config.U_H_correction_factor,
                 )
-                coefficient_data.to_hdf(output_path, key=store_group, mode="a")
+                coefficient_data.to_hdf(output_path, key=store_group, mode="a", format="table")
 
 
 def process_cp(
@@ -137,7 +156,7 @@ def process_cp(
     cfg: CpConfig,
     mesh: LnasGeometry,
     path_manager: CpPathManager,
-    extreme_params: ExtremeValuesParameters | None,
+    time_scale_factor: float,
 ):
     """Executes the pressure coefficient processing routine
 
@@ -148,10 +167,9 @@ def process_cp(
         cfg (CpConfig): Pressure coefficient configuration
         mesh (LnasGeometry): Geometry of the body
         path_manager (CpPathManager): Object to handle paths
-        extreme_params (ExtremeValuesParameters | None): Optional parameters for extreme values analysis
+        time_scale_factor (float): Factor for converting time scales from CST values
     """
-    cfg_hash = cfg.sha256()
-    timeseries_path = path_manager.get_cp_t_path(cfg_lbl=cfg_label, cfg_hash=cfg_hash)
+    timeseries_path = path_manager.get_timeseries_path(cfg_lbl=cfg_label)
     create_folders_for_file(timeseries_path)
 
     if timeseries_path.exists():
@@ -169,7 +187,7 @@ def process_cp(
         cp_config=cfg,
     )
 
-    grouped_data_path = path_manager.get_grouped_cp_path(cfg_lbl=cfg_label, cfg_hash=cfg_hash)
+    grouped_data_path = path_manager.get_grouped_timeseries_path(cfg_lbl=cfg_label)
 
     if grouped_data_path.exists():
         warnings.warn(
@@ -189,12 +207,11 @@ def process_cp(
     cp_stats = calculate_statistics_for_groups(
         grouped_data_path=grouped_data_path,
         statistics=cfg.statistics,
-        extreme_params=extreme_params,
+        time_scale_factor=time_scale_factor,
     )
+    stats_path = path_manager.get_stats_path(cfg_lbl=cfg_label)
+    cp_stats.to_hdf(path_or_buf=stats_path, key="stats", mode="w", index=False, format="table")
 
-    stats_path = path_manager.get_cp_stats_path(cfg_lbl=cfg_label, cfg_hash=cfg_hash)
-    cp_stats.to_hdf(path_or_buf=stats_path, key="cp_stats", mode="w", index=False)
-
-    vtp_path = path_manager.get_vtp_path(cfg_lbl=cfg_label, cfg_hash=cfg_hash)
+    vtp_path = path_manager.get_vtp_path(cfg_lbl=cfg_label)
     polydata = create_polydata_for_cell_data(data=cp_stats, mesh=mesh)
     write_polydata(vtp_path, polydata)