+++ cfdmod/use_cases/pressure/shape/Ce_config.py
@@ -3,16 +3,13 @@ from __future__ import annotations
 __all__ = ["CeConfig", "CeCaseConfig"]
 
 import pathlib
-from typing import Optional
 
-from pydantic import BaseModel, Field, field_validator, model_validator
+from pydantic import BaseModel, Field, field_validator
 
 from cfdmod.api.configs.hashable import HashableConfig
 from cfdmod.api.geometry.transformation_config import TransformationConfig
-from cfdmod.use_cases.pressure.extreme_values import ExtremeValuesParameters
+from cfdmod.use_cases.pressure.base_config import BasePressureCaseConfig, BasePressureConfig
 from cfdmod.use_cases.pressure.shape.zoning_config import ZoningConfig
-from cfdmod.use_cases.pressure.statistics import Statistics
-from cfdmod.use_cases.pressure.zoning.processing import ShapeVariables
 from cfdmod.utils import read_yaml
 
 
@@ -29,23 +26,13 @@ class ZoningBuilder(BaseModel):
         return zoning_cfg
 
 
-class CeConfig(HashableConfig):
+class CeConfig(HashableConfig, BasePressureConfig):
     """Configuration for shape coefficient"""
 
     zoning: ZoningConfig | ZoningBuilder = Field(
         ...,
         title="Zoning configuration",
         description="Zoning configuration with intervals information",
-    )
-    variables: list[ShapeVariables] = Field(
-        ...,
-        title="List of variables",
-        description="Define which variables will be calculated",
-    )
-    statistics: list[Statistics] = Field(
-        ...,
-        title="List of statistics",
-        description="List of statistics to calculate from shape coefficient signal",
     )
     sets: dict[str, list[str]] = Field(
         {}, title="Surface sets", description="Combine multiple surfaces into a set of surfaces"
@@ -78,25 +65,12 @@ class CeConfig(HashableConfig):
             raise Exception("Surfaces inside a set cannot be listed in zoning")
 
 
-class CeCaseConfig(BaseModel):
+class CeCaseConfig(BasePressureCaseConfig):
     shape_coefficient: dict[str, CeConfig] = Field(
         ...,
         title="Shape Coefficient configs",
         description="Dictionary of shape coefficient configurations",
     )
-    extreme_values: Optional[ExtremeValuesParameters] = Field(
-        None,
-        title="Extreme values parameter",
-        description="Parameters for performing extreme value analysis",
-    )
-
-    @model_validator(mode="after")
-    def check_extreme_values_params(self) -> CeCaseConfig:
-        full_stats = [s for v in self.shape_coefficient.values() for s in v.statistics]
-        if any(stats in full_stats for stats in ["xtr_min", "xtr_max"]):
-            if self.extreme_values is None:
-                raise ValueError("Extreme values parameters must be specified!")
-        return self
 
     @classmethod
     def from_file(cls, filename: pathlib.Path) -> CeCaseConfig: